fix(build): fetch the zfs header from the correct branch
autoreconf is always fetching the zfs header file from zfs repo
master branch. This make it incompatible to build on released
branches if there has been a change to the header file on the master.

Added script to determine the zfs branch from which the header
file should be fetched.
- If tagged build, fetch from the corresponding tag
- If master (replication), fetch from the zfs master
- If non-master, fetch from corresponding zfs branch if it exists
  or use the zfs master

Signed-off-by: kmova <kiran.mova@openebs.io>