feat(build): build docker container
This repository contains the complete code for istgt
and has a dependency to dynamically fetch a zfs header
file from openebs/zfs repo.

The previous commit(#81) tried to make the fetch automated
by using the current branch name as a key to identify the
dependent branch. However, not all scenarios got covered.
Besides having the image built in openebs/cstor repo - made
it more tricky to determine the current branch that needs
to be used as key.

Also, building in opeenbs/cstor repo kind of de-links the
commits from the images and depend on a manual trigger
by the cstor repo to build a docker image.

This PR addresses the following:
- Add code to build the docker images after a successful build.
- Put a place holder for specifying the dependent zfs branch

Signed-off-by: kmova <kiran.mova@openebs.io>