Reworked Docker image to suit style guidelines
The provided Docker image did not suit the Docker style guidelines as defined
by the Docker documentation [1]. This change greatly reduces the number of
layers which in effect reduces the image size by about 330MB [2]. Although this
size can differ a bit depending on the source.

This change also updates the base image to Ubuntu 16 instead of 14. I did this
because Ubuntu 14 will be end of life next year, while the Ubuntu 16 LTS will
only be EOL in 2021 [3].

 * [1] https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#minimize-the-number-of-layers
 * [2] https://imgur.com/a/SQg7ka4
 * [3] https://wiki.ubuntu.com/Releases

Signed-off-by: Alex van den Hoogen <alex3305@gmail.com>