fix(snaprebuild) when we are transitioning to
AFS mode and we get a snpashot request from the user,
there is a race that if snapshot command is received
after we decided to do AFS transition, we will miss or
have the inconsistant snapshot.

Here traget it self will send prepare snapshot comand to make
each replica ready to receive the snapshot command and then it will
send the command.

Signed-off-by: Pawan <pawan@mayadata.io>